third_party/blink/web_tests/platform/generic/fast/replaced/iframe-with-percentage-height-within-table-with-anonymous-table-cell-expected.txt

This test checks that iframe with percentage height within anonymous table cell has the correct height (strict mode).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS getHeight('iframe-100') is window.innerHeight
PASS successfullyParsed is true

TEST COMPLETE
